Update machine.adoc
Clarifying behavior of mstatus.SIE and mstatus.SPIE being read-only 0 when supervisor mode is not implemented.   

If this is not accurate, it would be illuminating to say that "SIE and SPIE might be writable but are ignored when supervisor mode is not implemented." However, this would be inconsistent with SPP must be read only 0 if supervisor mode is not implemented at the end of this section.

Signed-off-by: David Harris <74973295+davidharrishmc@users.noreply.github.com>